Call quits for the night
I know what I need to do here, but I also feel like I got hit
with a wave of inertia when it came to how bloated my
styling is. Its just so, so yucky.

I likely am directly overwriting the same elements multiple times,
making a mess of alignment, et cetera. I need to add a div class
which formats an h4 and a p to be the journal entries for both
right banner and main content. UGH. Hopefully, I'll have more
energy for this in the morning.